pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "ssdd-calculator"
version = "1.0.0"
authors = [
  { name="Javier Plaza Sánchez", email="javier.plaza3@alu.uclm.es" },
]
description = "Calculadora remota usando Apache Kafka"
requires-python = ">=3.10"
dependencies = ['zeroc-ice']
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
]
license = "GPL-3.0-only"

[project.scripts]
ssdd-calculator = "calculator.command_handlers:calculator"
